foileditors headers to include dir
