+++ include/foileditors/fwd/foileditorsfwd.hpp
@@ -1,54 +1,34 @@
 /****************************************************************************
-  
- Copyright (c) 2013, Hans Robeers
+
+ Copyright (c) 2014, Hans Robeers
  All rights reserved.
- 
+
  BSD 2-Clause License
- 
+
  Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
- 
+
    * Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer.
-   
+
    * Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution.
-   
+
  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
  COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
  (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
  HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
  ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- 
+
 ****************************************************************************/
 
-#ifndef THICKNESSCONTOURS_H
-#define THICKNESSCONTOURS_H
-
-#include "patheditor/fwd/patheditorfwd.hpp"
-#include "foillogic/fwd/foillogicfwd.hpp"
-
-#include <QGraphicsObject>
-#include <QThreadPool>
-
-using namespace patheditor;
+#ifndef FOILEDITORSFWD_HPP
+#define FOILEDITORSFWD_HPP
 
 namespace foileditors
 {
-    class ThicknessContours : public QGraphicsObject
-    {
-    public:
-        explicit ThicknessContours(foillogic::FoilCalculator* calculator, foillogic::Side::e side, QGraphicsItem *parent = 0);
-
-        virtual void paint(QPainter *painter, const QStyleOptionGraphicsItem *option, QWidget *widget);
-        virtual QRectF boundingRect() const;
-
-        virtual ~ThicknessContours() {}
-
-    private:
-        foillogic::Side::e _side;
-        bool _nextDetailed;
-
-        foillogic::FoilCalculator* _calculator;
-    };
+    class OutlineEditor;
+    class ProfileEditor;
+    class ThicknessEditor;
+    class FoilDataWidget;
 }
 
-#endif // THICKNESSCONTOURS_H
+#endif // FOILEDITORSFWD_HPP
